% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdverseEvents.R
\name{simulate_events}
\alias{simulate_events}
\title{Simulate many events
This function simulates multiple events based on the provided event parameters.}
\usage{
simulate_events(AEs, max_time, n_it = 100)
}
\arguments{
\item{AEs}{AE or a list of  of AE objects}

\item{max_time}{Maximum time to simulate events (in days)}
}
\value{
A list of event objects with simulated event times
}
\description{
Simulate many events
This function simulates multiple events based on the provided event parameters.
}
\examples{
AEs <- list(
 AE(10, 1, 10),
AE(10, 1, 10),
AE(10, 1, 10)
)
simulated_AEs <- simulate_events(AEs, 180)
}
\seealso{
\code{\link{AE}} for defining an event

\code{\link{simulate_event}} for simulating a single event
}
