% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdverseEvents.R
\name{simulate_events_}
\alias{simulate_events_}
\title{Simulate a single set of events
This function simulates multiple events based on the provided event parameters.}
\usage{
simulate_events_(AEs, max_time)
}
\arguments{
\item{one}{AE or a list of  of AE objects}
}
\description{
Simulate a single set of events
This function simulates multiple events based on the provided event parameters.
}
