% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenarios.R
\name{simulate_scenario}
\alias{simulate_scenario}
\title{Simulate a scenario}
\usage{
simulate_scenario(scenario, n_sim = n_sim)
}
\arguments{
\item{scenario}{A scenario object}

\item{n_sim}{Number of simulations to run}
}
\description{
This function simulates a scenario and saves the results to a json file.
}
